refactor(orders-table): move styled elements from OrdersTablePagination (#3896)
